backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
bcrypt
duckdb==1.4.4
pydantic
